renaming and move Input component, remove reset()
