+++ src/components/Form/Form.tsx
@@ -1,11 +1,11 @@
 import styles from "./Form.module.css";
 import { useForm, type SubmitHandler } from "react-hook-form";
-import type { IFormInput, TFormField, FormProps } from "./types";
+import type { IFormInput, TFormField, FormProps, TFormValues } from "./types";
 import { InputWrapper } from "../InputWrapper/InputWrapper";
 import { useEffect } from "react";
-import { Input } from "../Input/Input";
-import { formSchema, type TFormValues } from "../../constants";
+import { InputPersonalData } from "./ui/InputPersonalData/InputPersonalData";
 import { zodResolver } from "@hookform/resolvers/zod";
+import { formSchema } from "./lib/validation";
 
 export const Form = ({ labels, formData, setFormData }: FormProps) => {
   const {
@@ -21,11 +21,14 @@ export const Form = ({ labels, formData, setFormData }: FormProps) => {
   const onSubmit: SubmitHandler<IFormInput> = () => reset();
 
   useEffect(() => {
-    reset(formData);
-  }, [formData]);
+    console.log("first");
+    console.log(formData);
+    // reset(formData);
+  }, []);
 
   useEffect(() => {
     const subscription = watch((data) => {
+      console.log("hi");
       const hasValues = Object.values(data).some((value) => value);
       if (hasValues) setFormData(data);
     });
@@ -38,7 +41,7 @@ export const Form = ({ labels, formData, setFormData }: FormProps) => {
       {Object.entries(labels).map(([key, label], index) => {
         return (
           <InputWrapper key={index} labelTitle={label} errorMessage={errors[key as TFormField]?.message}>
-            <Input name={key as TFormField} configuration={register} />
+            <InputPersonalData name={key as TFormField} configuration={register} />
           </InputWrapper>
         );
       })}
