+++ src/components/Form/types.ts
@@ -1,4 +1,6 @@
 import type { FormHTMLAttributes } from "react";
+import type { formSchema } from "./lib/validation";
+import type { z } from "zod";
 
 export interface FormProps extends FormHTMLAttributes<HTMLFormElement> {
   labels: { [key: string]: string };
@@ -8,6 +10,7 @@ export interface FormProps extends FormHTMLAttributes<HTMLFormElement> {
 }
 
 export type TFormField = "firstName" | "lastName" | "familyName" | "birthDate" | "gender";
+
 export interface IFormInput {
   firstName: string;
   lastName: string;
@@ -15,3 +18,5 @@ export interface IFormInput {
   birthDate: string;
   gender: "male" | "female";
 }
+
+export type TFormValues = z.infer<typeof formSchema>;
