+++ src/components/Form/ui/InputPersonalData/types.ts
@@ -1,6 +1,6 @@
 import type { InputHTMLAttributes } from "react";
 import type { UseFormRegister } from "react-hook-form";
-import type { IFormInput, TFormField } from "../Form/types";
+import type { IFormInput, TFormField } from "../../types";
 
 export interface InputProps extends InputHTMLAttributes<HTMLLabelElement> {
   name: TFormField;
