+++ src/hooks/types.ts
@@ -1,7 +1,3 @@
 import type { IFormInput } from "../components/Form/types";
 
-export type TUseLocalStorage = [
-    Partial<IFormInput> | undefined,
-    (key: Partial<IFormInput>)=> void,
-    () => void,
-];
+export type TUseLocalStorage = [Partial<IFormInput> | undefined, (key: Partial<IFormInput>) => void, () => void];
